backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
PyJWT
python-dotenv
django-cors-headers
Pillow
drf-spectacular
drf-yasg
